[PLAT-7297] non-root install for yba installer works again
Summary:
Revamped how we handle "base" directories to enable yba-installer on non-root
installs. This allows us to easily change between /opt/yugabyte(ybactl) and
~/yugabyte(ybactl) depending on the install type.

From there, several improvements had to made to ensure the non-root install]
workflow implemented the same steps as the normal, root install workflow.

Last, all of the manage*.sh cron scripts needed some revamping to handle the
new directory layouts we implemented some time ago

This should fully cover thie install workflow, but we still need to handle
upgrade workflow

Test Plan:
Ran install -> clean multiple times. both with and without root
validated yba actually starts and is accessible after successful install

Reviewers: sanketh, muthu

Reviewed By: muthu

Subscribers: jenkins-bot, sanketh, muthu, yugaware

Differential Revision: https://phabricator.dev.yugabyte.com/D22803

+++ managed/yba-installer/cmd/prometheus.go
@@ -89,11 +89,7 @@ func (prom Prometheus) Install() {
 	if common.HasSudoAccess() {
 		userName := viper.GetString("service_username")
 		common.Chown(common.GetSoftwareRoot()+"/prometheus", userName, userName, true)
-
-	}
-
-	//Crontab based monitoring for non-root installs.
-	if !common.HasSudoAccess() {
+	} else {
 		prom.CreateCronJob()
 	}
 
@@ -111,13 +107,16 @@ func (prom Prometheus) Start() {
 		common.RunBash(common.Systemctl, []string{"status", "prometheus"})
 
 	} else {
-		bashCmd := fmt.Sprintf("%s %d %d %d %d %d > /dev/null 2>&1 &",
+		bashCmd := fmt.Sprintf("%s %s %s %d %d %d %d %d %s > /dev/null 2>&1 &",
 			prom.cronScript,
+			common.GetSoftwareRoot(),
+			common.GetDataRoot(),
 			viper.GetInt("prometheus.port"),
 			viper.GetInt("prometheus.maxConcurrency"),
 			viper.GetInt("prometheus.maxSamples"),
 			viper.GetInt("prometheus.timeout"),
 			viper.GetInt("prometheus.restartSeconds"),
+			prom.version,
 		)
 		command1 := "bash"
 		arg1 := []string{"-c", bashCmd}
@@ -154,8 +153,11 @@ func (prom Prometheus) Stop() {
 		if strings.TrimSuffix(string(out0), "\n") != "" {
 			pids := strings.Split(string(out0), "\n")
 			for _, pid := range pids {
-				argStop := []string{"-c", "kill -9 " + strings.TrimSuffix(pid, "\n")}
-				common.RunBash(commandCheck0, argStop)
+				log.Debug("kill prometheus pid: " + pid)
+				if strings.TrimSuffix(pid, "\n") != "" {
+					argStop := []string{"-c", "kill -9 " + strings.TrimSuffix(pid, "\n")}
+					common.RunBash(commandCheck0, argStop)
+				}
 			}
 		}
 	}
@@ -180,6 +182,7 @@ func (prom Prometheus) Restart() {
 
 // Uninstall stops prometheus and optionally removes all data.
 func (prom Prometheus) Uninstall(removeData bool) {
+	log.Info("Uninstalling prometheus")
 	prom.Stop()
 
 	if common.HasSudoAccess() {
@@ -198,7 +201,8 @@ func (prom Prometheus) Uninstall(removeData bool) {
 	if removeData {
 		err := common.RemoveAll(prom.prometheusDirectories.DataDir)
 		if err != nil {
-			log.Info(fmt.Sprintf("Error %s removing data dir %s.", err.Error(), prom.prometheusDirectories.DataDir))
+			log.Info(fmt.Sprintf("Error %s removing data dir %s.", err.Error(),
+				prom.prometheusDirectories.DataDir))
 		}
 	}
 }
@@ -228,7 +232,8 @@ func (prom Prometheus) Upgrade() {
 func (prom Prometheus) moveAndExtractPrometheusPackage() {
 
 	srcPath := fmt.Sprintf(
-		"%s/third-party/prometheus-%s.linux-amd64.tar.gz", common.GetInstallerSoftwareDir(), prom.version)
+		"%s/third-party/prometheus-%s.linux-amd64.tar.gz", common.GetInstallerSoftwareDir(),
+		prom.version)
 	dstPath := fmt.Sprintf(
 		"%s/packages/prometheus-%s.linux-amd64.tar.gz", common.GetInstallerSoftwareDir(), prom.version)
 
@@ -343,13 +348,17 @@ func (prom Prometheus) Status() common.Status {
 // CreateCronJob creates the cron job for managing prometheus with cron script in non-root.
 func (prom Prometheus) CreateCronJob() {
 	bashCmd := fmt.Sprintf(
-		"(crontab -l 2>/dev/null; echo \"@reboot %s %s %s %s %s %s \") | sort - | uniq - | crontab - ",
+		"(crontab -l 2>/dev/null; echo \"@reboot %s %s %s %s %s %s %s %s %s \") | sort - | uniq - | "+
+			"crontab - ",
 		prom.cronScript,
+		common.GetSoftwareRoot(),
+		common.GetDataRoot(),
 		config.GetYamlPathData("prometheus.port"),
 		config.GetYamlPathData("prometheus.maxConcurrency"),
 		config.GetYamlPathData("prometheus.maxSamples"),
 		config.GetYamlPathData("prometheus.timeout"),
 		config.GetYamlPathData("prometheus.restartSeconds"),
+		prom.version,
 	)
 	common.RunBash("bash", []string{"-c", bashCmd})
 }
